Refactor OracleEnhanced::Connection responsibilities
Split the base Connection class into a thin dispatcher and a separate
DatabaseDescription module so each piece has a single responsibility.

* Connection.create now delegates to a registered connection_class
  instead of branching on ORACLE_ENHANCED_CONNECTION. The constant is
  kept because database_statements.rb and the spec suite still read it.
* describe / _select_one / _select_value / _oracle_downcase move to a
  new OracleEnhanced::DatabaseDescription module included by both
  OCIConnection and JDBCConnection. This removes the awkward
  "def describe(name); super; end" shims that only existed to expose
  the base class's private method.
